updates + recoding figures with updated final code from JC
Note: I added comments/section headers to the final script to more easily locate the code for each figure.

I also edited the herps data preparation near the beginning, because the wild species were being lost when creating herpswild and herpslpi, which was making the chi-squared tests impossible (we only had 1 group variable, which was c-lpi).